pyproject.toml
[project]
name = "tika-proxy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.12",
    "fastapi>=0.115.8",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
